Bug Fix: Remove all `forceSvg` specific logic (#6244)
## Motivation for features / changes
`forceSvg` being stored to localStorage predates our feature allowing
any feature flag to be stored to localStorage. Unfortunately these two
features work quite poorly together.

The `forceSvg` flag has turned into a kind of virus because it writes
itself to localStorage when it appears in the url. This means that users
who having it enabled sharing a link to another user will cause the
setting to spread.

Now that there is a new way of enabling this feature permanently for
users, I am removing the older method of persisting the flag.

## Technical description of changes


## Screenshots of UI changes
None

## Detailed steps to verify changes work correctly (as executed by you)
1) Start tensorboard
2) Navigate to localhost:6006?showFlags
3) Ensure `forceSvg` is disabled
4) Navigate to localhost:6006?showFlags&forceSvg=true
5) Ensure `forceSvg` is enabled
6) Navigate to localhost:6006?showFlags
7) Ensure `forceSvg` is still disabled

## Alternate designs / implementations considered
I could have just updated the `force_svg_data_source` to stop storing
force_svg differently and potentially migrate users to store `forceSvg`
using the same persistence as other feature flags. However, we do not
believe there are a lot of users that actually need the feature and thus
we would prefer to force the small number of users that need this
feature to re-enable it while rescuing unintentionally effected users.