Feature: Update the showFlags feature to allow filtering (#6245)
## Motivation for features / changes
As we add more feature flags it's starting to get hard to find them.
This is particularly true when trying to help someone else find a
feature flag in the modal.

The `enableShowFlags` feature was a weird edge case that wasn't being
fully utilized. Now it supports string values. When a string value is
supplied it filters the feature flags shown.

## Technical description of changes
I changed the name and value of `enableShowFlags: boolean` to
`showFlags: string`.
I then updated the `feature_flag_page_container` to filter the feature
flags rendered based on the value of the `showFlags` feature.

## Screenshots of UI changes
Modal With a Filter

![image](https://user-images.githubusercontent.com/78179109/225770098-5460517e-8946-42f2-8f6a-b26980b63ffc.png)

Modal without a filter

![image](https://user-images.githubusercontent.com/78179109/225770279-034c0eb7-a566-4376-b265-39d6adae4b9d.png)

## Detailed steps to verify changes work correctly (as executed by you)
1) Start tensorboard
2) Navigate to localhost:6006
3) Assert the feature flags modal does not appear
4) Navigate to localhost:6006?showFlags
5) Assert the feature flags modal appears
6) Assert the no filter is applied
7) Navigate to localhost:6006?showFlags=enable
8) Assert the feature flags modal appears and has a filter applied

## Alternate designs / implementations considered
Adding the filter could automatically toggle a feature?
Adding an input field to filter things would make sense...