Bumps google-auth-oauthlib dependency restriction to support up to version 1.0 (#6246)
* Motivation for features / changes
As requested and noted in #6197, we can bump this restriction, which is
currently affecting compatibility of TB with dependency requirements
from other packages. After the OOB auth flow migration, this should be
possible.

* Technical description of changes
Just bump the restriction for `google-auth-oauthlib` in our
`requirements.txt` file, to support newer versions.

* Detailed steps to verify changes work correctly (as executed by you)
Executed the auth flows (both the console-only one, and the one that
uses a browser in the same machine), and validated that they work as
before with the newer version.