Column Customization: Persist Column Settings (#6253)
* Motivation for features / changes
We now have a way for users to customize which columns they would like
to appear in their data table and in what order. However, we currently
revert back to default on every page load. This change stores their
selection in localStorage so that it persists a crossed sessions.

* Technical description of changes
This just follows the same structure as all persisted settings in OSS.