Cleanup: Remove remnants of feature flag route. (#6261)
In go/tbpr/5760 we introduced the FLAGS route. The idea was to build a
new page within TensorBoard, devoted to setting internal feature flags.

We pivoted and instead of having a FLAGS route we have a flags dialog
that pops up with the showFlags query parameter, regardless of route.

This PR removes the remants of the FLAGS route. it also renames the
`feature_flag_page*.ts` files and `FeatureFlagPage*` TS objects to
`feature_flag_dialog*.ts` and `FeatureFlagDialog*`, respectively.

To Test:
Made sure TensorBoard builds and runs and tests pass.
Opened TensorBoard with showFlags query parameter and verified the
dialog appears and works.