Add mechanism to prevent state to be persisted in local storage. (#6267)
For Google-internal versions of TensorBoard, there are some scenarios
where we do not want to read or write user settings to local storage.
This PR adds a mechanism to facilitate turning persistence off, if
desired.

The PR introduces persistent_settings State with one property,
shouldPersistSettings, and one selector, getShouldPersistSettings. In
this codebase the value of this is always true. Google-internal versions
of TensorBoard, however, may have their own reducers that set this
property to false.

The PR also modifies persistent_settings effects to read the value of
shouldPersistSettings before determining whether to read and write to
the persistent data source. The effects now wait until the first
navigating() event before attempting to read from the data source, in
order to give other layers a small amount of time to set the
shouldPersistSettings state before the first read attempt. There's
nothing particularly special about using the navigating() event for
this, other than it is fired late enough for our current use cases. We
could consider using a different, later, action in the future.

Googlers, see how one internal TensorBoard overrides this property in
cl/518939981.