Fix safari chart rendering issues. (#6303)
* Motivation for features / changes

I found a couple of problems with scalar line chart rendering in Safari
while investigating
https://github.com/tensorflow/tensorboard/issues/6280.

* Safari 16.4 introduced limited OffscreenCanvas support. Notably it
does not include 'webgl2' support. We have to be a bit more strict with
our OffscreenCanvas feature detection to check for 'webgl2' support.
* Scalar line charts still would not render in either canvas or svg mode
until the charts were resized manually. We have to adjust our CSS in
order to convince WebKit to render the line chart with the space
available to it. We use 'flex' for this.

* Testing

  I tested in both Safari and Chrome that Canvas/Three.js and SVG
rendering work (using forceSVG=false and forceSVG=true, respectively).

  I imported the changes internally and ran internal presubmits.

  I did not write any more tests because (1) utils.ts is currently not
covered by tests and (2) this is a Safari problem which only has best
effort support.