Column Customization: Use dataTableColumnEdited action when dragging columns in data table (#6291)
* Motivation for features / changes
Before this change, dragging the header in the data table called the
dataTableColumnDrag action. This action relied on the global
rangeSelectionEnabled boolean to decide if it edits the range of single
selection headers. This boolean no longer is guaranteed to represent the
table which is being edited which can lead to a bug where the single
selection table shows the range selection headers and vise versa.

* Technical description of changes
This moves the decision on which columns to edit to the
ScalarCardDataTable component. It also reuses the dataTableColumnEdited
action which is used by the ScalarColumnEditor component.

* Screenshots of UI changes
This is the broken state which is possible before this change. Note that
the single selection has the range columns.
<img width="900" alt="Screenshot 2023-04-05 at 11 52 24 AM"
src="https://user-images.githubusercontent.com/8672809/230177414-bdadc95c-e416-4cce-abcf-6cca55c9dd61.png">