HParams: Port runs_data_source fetchHparamsMetadata to oss (#6318)
## Motivation for features / changes
As part of the effort to surface hparams data in the time series
dashboard we need to start actually fetching it. The logic existed
internally but for some reason was not previously available in OSS.

## Technical description of changes
I just copied the internal code to the oss implementation and made a few
small adjustments to imports

## Screenshots of UI changes
None

## Detailed steps to verify changes work correctly (as executed by you)
Tests should pass
Patch #6317, enable the feature flag, and ensure the hparams appear in
the runs table

## Alternate designs / implementations considered
Make a more detailed stub?