HParams: Add new hparams selector factories (#6323)
## Motivation for features / changes
As part of the effort to bring hparams to the time series dashboard a
handful of new selectors need to be created. Rather than further the
practice of creating selectors with props ([which is
deprecated](https://ngrx.io/guide/migration/v12)) I am choosing to
create them using the recommended solution of a selector factory.
Unfortunately selector factories don't play nicely with selectors
dependent selectors WITH props. There are a handful of hacky solutions
to this problem but I am opting to create new dependent selectors using
the factory pattern.

## Technical description of changes
I extracted the result functions from the existing selectors, then
created new factories which call them with an additional closured
parameter.

## Screenshots of UI changes
None

## Detailed steps to verify changes work correctly (as executed by you)
Tests should pass

## Alternate designs / implementations considered
I could have just wrapped the existing selectors and then called them
passing in the global state instead. I chose this approach instead with
the hope of removing the old selectors in a later pr.