Column Customization: Open Editor to the expected tab (#6305)
## Motivation for features / changes
When the Column Editor is opened from the scalar card it opens to
whatever the default or last selected tab is. However, it would be
expected to open to the mode that the card used to open it is in(Single
vs Range). This makes that happen

## Technical description of changes
This PR creates a new value in the metricsState(tableEditorSelectedTab)
which reflects the selected tab in the table editor. It then pulls that
value into the editor and uses it to select the tab. It is kept up to
date with calls to tableEditorTabChanged (a new action created for this
purpose) when the tab is changed inside the controller.

Lastly it adds a new prop to the metricsSlideoutMenuOpened action which
contains the current mode of the chart which dispatched the action. This
mode is used to update the new tableEditorSelectedTab value.

## Screenshots of UI changes
![2023-04-13_23-47-03
(1)](https://user-images.githubusercontent.com/8672809/231965710-9b124054-cf47-4394-9789-18894e06574c.gif)