HParams: Add new selector factory to get runs from experiment ids (#6360)
## Motivation for features / changes
As part of the effort to bring hparams to the time series dashboard a
handful of new selectors need to be created. In particular I will be
creating a new one to handle runs filtering by moving the logic out of
the runs_table_container into the selector layer.
This will ensure filtered runs are consistent across the dashboard.

## Technical description of changes
This new selector is not meaningfully different from the existing one
but the factory allows it to be called with an encapsulated experiment
id.