HParam UI: reset ColumnHeaders if using obsolete interface (#6358)
## Motivation for features / changes
In #6346 we updated the ColumnHeader interface. However, objects using
this interface are stored in localStorage. This checks localStorage to
ensure it has the new interface. If it does not then it ignores those
values.

## Technical description of changes
I am simply checking if the first value has the name property to ensure
the values are valid. I could do a more thorough check but it seems
unnecessary.

## Detailed steps to verify changes work correctly (as executed by you)
It worked for me when running locally.