Use maximum step as default for single step selection. (#6372)
## Motivation for features / changes

We are changing single step selection to default to the maximum step. We
think this is more useful than making the default the minimum step.

This also means that when step selection changes from single selection
to range selection, the minimum value of the chart is added as the start
of the range and the previous single step now becomes the end of the
range. Similarly, when step selection changes from range selection to
single selection, the new single selection is the previous maximum value
of the range.

## Technical description of changes

Where default value is set to "min" step, instead set it to the "max"
step. Where range selection is enabled, ensure the new start of the
range is the "min" step.

I have three internal CLs that accompagny this change.
* cl/528745120 adjusts or comments out code to allow the import to
succeed. It will be submitted prior to merging this PR.
* cl/528745440 is a set of harmless screenshot-updates. It will be
patched into the import CL and submitted at the same time as the import.
* cl/528745657 is a set of changes to uncomment out the code from the
first CL and make remaining adjustments. It will be submitted after the
import.

## Detailed steps to verify changes work correctly (as executed by you)

* I tested it manually quite a bit.
* I imported it into the internal repo and ensured that end-to-end tests
pass (after adjusting for the change in behavior).