HParam UI: Change SelectedStepRunData to use name (#6362)
## Motivation for features / changes
The eventual goal is to allow dynamic HParam columns in the data table.
To do this we introduced the name and displayName attributes to the
ColumnHeader in #6346. The name attribute is now going to be the unique
identifier for a column instead of the type enum.

## Technical description of changes
The SelectedStepRunData is used to hold the data for the table for a
specific run. Previously this was an object with an optional property
for all values of the ColumnHeaderType enum. This will no longer work as
these values need to be more dynamic to allow for all possible Hparam
values. This PR changes that object to add an array which holds all the
data points associated with the name of the column which they correspond
to.

Unfortunately this breaks the sorting as the array of
SelectedStepRunData which is passed to the DataTable is sorted based on
which header type the user chose to sort by. So that had to be updated
to use the name of the column which the user chose to sort by.

## Alternate designs / implementations considered (or N/A)
I thought about adding a new field to the SelectedStepRunData object
called HParam value and keeping all the optional attributes. This was a
big headache for ordering and sorting.