HParam UI: Move types from scalar_card_types to data_table/types (#6378)
## Motivation for features / changes
We plan to use the data_table widget in place of the runs table. These
types will be needed there so it makes more sense for them to be part of
the data_table widget instead of the scalar_card.

I also changed the name of the SelectedStepRunData to a much more
generic TableData. As this will be the object used for each row in the
runs table as well which has nothing to do with the selected step.