Hparams: Use DataProvider.list_hyperparameters() to generate HparamInfos (#6391)
Use the result from a DataProvider.list_hyperparameters() call to
generate HparamInfos for operations in the Hparams plugin. Generate the
HparamInfo name, display_name, data_type, and domain from each
Hyperparameter.

We use the same two-step strategy for generating these as is done with
tensor-based hparams. The tensor-based hparams has one operation,
`hparams_metadata()` for retrieving the tensor data, and then passes the
result to `experiment_metadata()` to generate the HparamInfos and
MetricInfos. Similarly, we add an `hparams_from_data_provider()`
operation for retrieving the hyperparameter data and we pass the result
to `experiment_metadata()`.

Note: We don't yet generate `MetricInfos`. We don't yet have enough
information from the data provider.

Note: We don't yet merge the new HparamInfos with the tensor-based
HparamInfos.