Bug Fix: update hparams plugin to generate domains for boolean hparams (#6393)
## Motivation for features / changes
This has been a bug for a while that we only noticed now that hparams
are being used to filter runs in the time series dashboard.

Googlers see https://chat.google.com/room/AAAA03izhrk/VyyPgojcNvY for
context on how this was discovered

## Technical description of changes
Boolean HParams were not having a domain set. This lead to the ui
treating them as intervals, however, this lead to incorrect filter
conditions. We currently treat all non number values being filtered this
way as not matching (the alternative would be to treat them all as being
true which also seems wrong). This leads to all runs with a value for a
boolean hparam being filtered out.

## Screenshots of UI changes (or N/A)

Googlers see cl/532932348

## Alternate designs / implementations considered (or N/A)
This could have been done on the client but ideally the bug would be
fixed in other places where the api is being used.