HParam UI: Replace RunsTableComponent with DataTable widget behind flag (#6394)
## Motivation for features / changes
We are adding lots of functionality to the scalar data table and the
runs table. For consistency and reuse we are replacing the table in the
runs table with a DataTable widget. This is the first step towards that
change.

## Technical description of changes
Besides the data all the values which are passed in are hard coded into
the runs_table_container. This is done so that we have a data table to
work with quickly and unblock some other work. These values will be set
up in ngrx in future PRs.

## Screenshots of UI changes (or N/A)
We have a plan to restyle this table for the Runs Table. However, that
will come later. For now it is just the standard data table which looks
like this.
<img width="775" alt="Screenshot 2023-05-17 at 4 13 21 PM"
src="https://github.com/tensorflow/tensorboard/assets/8672809/57cb2919-c046-40e0-8057-b753fac87176">

## Alternate designs / implementations considered (or N/A)
We considered leaving the runs table and adding all features to both.