Continue moving to use name in SortingInfo (#6400)
## Motivation for features / changes
For the HParam work that is being done we are now using the name field
as the unique identifier for columns. Therefore the SortingInfo
interface needs to use name and not header. However, making the hard
swap to using name would break sync so it is happening in stages. The
name switch to using names happened in #6362. However, we had to make
name optional for sync reasons. Googlers see cl/534097086 to see where
name was added. Now this PR makes header optional. I will then make an
internal change to remove it. Then I can finally submit a final PR which
removes the property entirely.