HParams: Update the runs table to read columns from redux state (#6404)
## Motivation for features / changes
As part the effort to add hparams to time series we have updated the
runs table to use the data table in #6394. Now I am updating the runs
data table to read its columns from the redux state.

## Screenshots of UI changes (or N/A)
N/A (but there will be soon!)

## Alternate designs / implementations considered (or N/A)
We could have reused the existing sorting property in redux and added
keys to it, but I prefered to create a seperate one and remove the
existing property when we remove the old runs table.