HParam: Add a new selector to create columns for hparams (#6406)
## Motivation for features / changes
As part of the upcoming hparams in timeseries feature we want to be able
to display hparams as columns in both the scalar card data table and the
runs table. Because we do not want all the hparams displayed all the
time (there may be too many) we are building an interface for selecting
columns. Towards that goal I have created this selector which finds all
the hparams and generates a column header for them.

## Screenshots of UI changes (or N/A)
N/A