HParams: project header cell components (#6422)
## Motivation for features / changes
This refactoring allows more customization of the data table which will
be used by both the runs table and scalar card. This allows the
parents(like the runs table) to add functionality to the header
component such as the color pallet and select all checkbox.

## Technical description of changes
This change pulls out the header cells from the DataTable and allows the
parent component to project them back down. We add a HeaderCellComponent
which is used keep consistency and allow the DataTable to still keep the
logic which should be on all tables such as dragging and sorting. This
is done by using event emitters in the HeaderCellComponent which the
DataTable subscribes to.

## Screenshots of UI changes (or N/A)
The Scalar table has no changes.
<img width="358" alt="Screenshot 2023-06-07 at 9 38 21 PM"
src="https://github.com/tensorflow/tensorboard/assets/8672809/7afd8a70-bfec-4707-a15c-f6abc28ec533">

This does break the runs data table which is still behind a flag. This
will be fixed in a follow up.
<img width="420" alt="Screenshot 2023-06-07 at 9 39 40 PM"
src="https://github.com/tensorflow/tensorboard/assets/8672809/22f856aa-1921-442d-9342-f423f4db5b6e">

## Detailed steps to verify changes work correctly (as executed by you)
Ran it and clicked a round a lot to try and break it.

## Alternate designs / implementations considered (or N/A)
We considered a more config based solution where we add lots of options
for functionality in the header objects being passed down to the table.