Change Hyperparameter to a dataclass. (#6426)
Change Hyperparameter to a dataclass so we can get special methods like
`__repr__` and `__eq__`. It also simplifies the code. We can do this now that we
only target versions of python > 3.7.

This also requires typing the Hyperparameter class in order for the
dataclass to be properly generated. I therefore considered also typing
the list_hyperparameters() function but there is mixed value since the
py_library() rule does not run a pytype checker.

Note that we still target Python 3.8 so we must use `Union` instead of
`|` in order to list `None` as an option.