HParams: project table content components (#6427)
## Motivation for features / changes
This is part of the effort to use the data table widget in the Runs
table which will allow for lots of code reuse as we add HParam
functionality to both tables. To allow more customization in each table
this PR pulls all the table content out of the data_table_component and
allows users to project it using the new content_row_component and
content_cell_component.

## Technical description of changes
Created new content_row_component and content_cell_component. The row is
extremely basic and basically just wraps the content inside a table-row.
The content_cell_component take all the logic for formatting each cell.
To customize the content users of this component can simply pass in no
datum or a ColumnHeaderType that does not have a specified way to format
in the getFormattedDataForColumn function and then put the custom
content inside the cell as a child it will be projected in. This is
currently shown in the way the ScalarCardDataTable handles the Color
column.

## Screenshots of UI changes (or N/A)
This should result in no changes to the Scalar Data Table. However, the
Runs Data Table is now empty as it has not implemented this new
DataTable structure. It is behind a flag though and will be implemented
in an upcoming PR.

<img width="398" alt="Screenshot 2023-06-09 at 2 24 42 PM"
src="https://github.com/tensorflow/tensorboard/assets/8672809/28002f09-bc93-45a1-99d5-752e4f71b2ed">

## Detailed steps to verify changes work correctly (as executed by you)
Ran it and clicked around a lot to try breaking.

## Alternate designs / implementations considered (or N/A)
We considered some way to allow the data table to be highly configured
instead of projecting the content.
I also considered not creating the content_row_component class and
requiring parents to add their own div with class=table-row.