HParams: project color header (#6434)
## Motivation for features / changes
In #6422 we started projecting the header components into the data table
widget. However, the content was still not projected and the color
column was hard coded in. Therefore we left a hard coded color header.
The content started being projected in #6427. Therefore we can stop
hardcoding this color column. This PR removes the hardcoded header
column and starts adding it from the ScalarCardDataTable

## Technical description of changes
The color header cannot be sorted, dismissed, or dragged. There was some
logic in place for this. However, it was incomplete. This completes that
logic and uses it for the color header. We also add tests for that logic
which were not in place before.

Note: I did some test clean up in the HeaderCell

## Screenshot of changes
It still works.
<img width="148" alt="Screenshot 2023-06-15 at 5 41 33 AM"
src="https://github.com/tensorflow/tensorboard/assets/8672809/0fe6e62e-f79a-46a8-bf60-cad3d87a28d4">
ComponentTests for consistency.

---------

Co-authored-by: Riley Jones <78179109+rileyajones@users.noreply.github.com>