HParam: Implement new Data Table projection architecture in Runs Table (#6438)
## Motivation for features / changes
We preparation for adding HParam functionality to both the Scalar Table
and the Runs Table we have refactored the Data Table to be more
flexible. This refactoring was done in #6422 and #6427 and was only
implemented in the Scalar Table because the Runs Table only uses the
Data Table behind a flag. This PR implements the new Data Table
Structure in the Runs Table.

## Technical description of changes
Basically just followed the same structure as the Scalar Table.

## Screenshots of UI changes (or N/A)
Runs Table specific styling still needs to be done. It currently looks
like this.
<img width="392" alt="Screenshot 2023-06-15 at 7 21 46 PM"
src="https://github.com/tensorflow/tensorboard/assets/8672809/2dd13003-ad93-4b66-89eb-579a497d6708">


We still need to add a column for check boxes and add styling.