Remove requirement that metricInfos are present to show Hparams main view. (#6452)
## Motivation for features / changes

We are building some new hparams-related functionality internally. It is
early stages and the data we return to the UI does not contain metrics.
The hparams dashboard, however, requires `metricInfos` to be non-empty
in order to show the main view of the dashboard. We want to show the
main view despite not having metric data.

More generally, though, it doesn't seem like metric data should be a
requirement to show hparam dashboard information. The hparams fields can
still be valuable on their own.

## Technical description of changes

Remove a UI-level check for non-empty metricInfos.