HParams: Sort Runs Table Data (#6459)
## Motivation for features / changes
The RunsDataTable has an affordance for sorting and the plumbing is
there to set the sorting info. However, the data was never sorted. This
PR sorts the data based on that sorting info which we have stored in
ngrx.

## Technical description of changes
I simply added another Obvservable which watches allRunsTableData$ and
sortingInfo$ and sorts. Then I passed that sorted observable to the
RunsDataTable.

## Screenshots of UI changes (or N/A)

## Detailed steps to verify changes work correctly (as executed by you)
I really only tested this with the run column. Then I wrote test for the
rest.

## Alternate designs / implementations considered (or N/A)