HParams: Order RunsDataTable Columns on drag (#6464)
## Motivation for features / changes
The dragging functionality was all in place except it never fired an
action to actually change the order. This PR creates the reordering
action/reducer and hooks up the orderColumns event with that action.

## Technical description of changes
As in the Scalar Table be re-ordering we actually just replace the
entire list of headers with a new one. I am not a huge fan of this. I
would like to change how this works sometime but that is beyond the
scope of this PR.

## Screenshots of UI changes (or N/A)
![2023-06-27_15-21-09
(1)](https://github.com/tensorflow/tensorboard/assets/8672809/9df1e85d-bb74-42ab-8282-0a28ff8137e0)

## Alternate designs / implementations considered (or N/A)
I considered refactoring the orderColumns events to just return source
and destination indexes. However, I decided that was too big a change
ATM. I think it should be done sometime in the future.