HParams: Separate data table controls (#6463)
## Motivation for features / changes
We have realized we need more direct control over which columns have
which controls. This change moves those controls to a set of optional
booleans in the ColumnHeader interface and utilizes them instead of a
general "ControlsEnabled" boolean which we used before. Functionally the
only change here should be that the runs column cannot be removed
anymore. However, this more direct control will help a lot as other
columns are added.

It also fixes a bug where header which could not sort were still
dispatching the sorting action.

## Technical description of changes
Added these as optional booleans and started using them instead of the
ControlsEnabled boolean. Since these headers are stored in localStorage
already for the Scalar Table I also had to go set them as a default in
the persistent settings data source.

## Screenshots of UI changes (or N/A)
Note the lack of a remove button:

![run_with_remove](https://github.com/tensorflow/tensorboard/assets/8672809/1b40d077-afad-40ed-8259-404ac454ec28)

## Detailed steps to verify changes work correctly (as executed by you)
I had old Single and Range headers stored in localStorage. They updated
properly now and work great. I messed with the runs table too. It all
seems to work. Except I found that ordering columns was never actually
hooked up. I will fix that in a follow up.

## Alternate designs / implementations considered (or N/A)
Considered just splitting the controlsEnabled prop to multiple
props...This seemed cleaner.