HParams: Allow an input to force use of legacy table (#6477)
## Motivation for features / changes
We will soon be flipping the flag to enable the RunsDataTable. However,
this new table does not work on the hosted experiments list page. This
allows that page(which is internal only) to force the runs table to use
the old table so we do not break that feature.

## Detailed steps to verify changes work correctly (as executed by you)
I ran this code internally first and used it to fix the problem. Then
copied it out to OSS.