HParams: Implement selection checkbox double click (#6481)
## Motivation for features / changes
We will soon be replacing the RunsTableComponent with the RunsDataTable.
In an effort to reach feature parity with the current runs table this PR
implements the dblClick functionality of the selected check boxes. When
the checkboxes are double clicked it has the effect of selecting this
and only this run.

## Technical description of changes
I implemented the dblclick logic differently than the
RunsTableComponent. I utilized the MouseEvent details property to decide
if a click was a click or a dblclick. This stops the click from firing
both the onSelectionToggle and the onSelectionDblClick events for a
single click.

## Screenshots of UI changes (or N/A)
![2023-07-05_16-57-38
(3)](https://github.com/tensorflow/tensorboard/assets/8672809/ea29d54e-5bb4-437d-9cf5-a09142f691fe)