Hparams: Change some handling/generation of hparams with discrete domains. (#6489)
This addresses several issues and cleanup related to how we generate and
handle hparams with discrete domains:

1. From the backend, always return the list of discrete domain values
for a string hparam. Previously we were returning no discrete domain
values when the number of values exceeded 10 but this was breaking
assumptions in the Angular hparam logic.
2. In the polymer UI, generate a "regexp" filter for discrete string
hparams with greater than 10 values.
3. Fix regexp filter display logic in the polymer UI. We were never
successfully showing the regexp filter. The check
`[[hparam.filter.regexp]]` was incorrect since it would evaluate to
`false` when `regexp` is the empty string. Instead we must check that
`hparam.filter.regex !== undefined`.
4. Simplify the logic for generating discrete domain filters in the
polymer UI - we just need to do it in a single place now that we can
assume that the backend will always return the list of discrete domain
values (thanks to item (1) and to previous work done in go/tbpr/6393)