Hparams: Add differs field. (#6500)
Add a `differs` field at both the DataProvider level (in Hyperparameter)
and at the hparams API level (in HParamInfo).

Also write logic to consume it in the hparams dashboard by sorting the
hparams that differ to the top.

There are, unfortunately, no tests in the repo for UI changes so I had
to test the changes manually. I tested that an experiment with < 5
hparams renders all hparams and selects all hparams. I tested that an
internal experiment (from an internal data provider implementation) that
has > 5 hparams sorts hparams with `differs===true` to the top and that
the first five are selected. I checked that attempts to set the order
and to set filters succeed.