Hparams: Treat Nan/Infinity numeric hparam values as unset values. (#6496)
## Motivation for features / changes

There are cases where users might log hparam numeric values that are NaN
or Infinity. Recall, we use proto's Value message for logging and
storing hparam values. While processing these values in the python layer
of the Hparams plugin, we may run these Value message through
json_format.MessageToJson, at which point the json_format library raises
an error.

The limitation is documented at
https://protobuf.dev/reference/protobuf/google.protobuf/#value.
"[A]ttempting to serialize NaN or Infinity results in error."

## Technical description of changes

We now will try to avoid calling MessageToJson on these problematic
hparam values. We instead chose to treat these as "unset" values --
equivalent to having not set an hparam value at all. It will mean the
values will not contribute to the calculation of discrete domains. And
it means that the values will appear blank in the hparams UI rather than
appearing as "NaN".

## Alternate designs / implementations considered (or N/A)

Given infinite amount of time:
* Fix json_format to handle NaN and Infinity :O.
* Change the hparams api.proto to use something other than proto's Value
for representing hparam values.

Other reasonable options:
* Change the NaN numeric value to "NaN" string. This was my first choice
but it unfortunately leads to other complications - for instance the
plugin will raise errors when it encounters a "NaN" string for an hparam
that is supposed to have a numeric value.